Add menu item for running logical query
